+++ Source/Architecture/UI/Button.cpp
@@ -33,17 +33,17 @@ void Button::setSelected(bool s) noexcept
 	selected = s;
 }
 
-int Button::getPosX() const noexcept
+float Button::getPosX() const noexcept
 {
 	return pos_x;
 }
 
-int Button::getPosY() const noexcept
+float Button::getPosY() const noexcept
 {
 	return pos_y;
 }
 
-void Button::setPos(int x, int y) noexcept
+void Button::setPos(float x, float y) noexcept
 {
 	pos_x = x;
 	pos_y = y;
